blog exercise 4.15 - 4.19

+++ part4/controllers/blogs.js
@@ -1,15 +1,25 @@
 const blogsRouter = require("express").Router()
 const Blog = require("../models/blog")
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+
+const getTokenFrom = request => {
+  const authorization = request.get('authorization')
+  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
+    return authorization.substring(7)
+  }
+  return null
+}
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog
-    .find({})
+    .find({}).populate('user', {username: 1, name: 1})
   response.json(blogs)
 })
 
 blogsRouter.get('/:id', async (request, response) => {
   const blog = await Blog.findById(request.params.id)
-  if(blog){
+  if (blog) {
     response.json(blog.toJSON())
   } else {
     response.status(404).end()
@@ -18,13 +28,24 @@ blogsRouter.get('/:id', async (request, response) => {
 
 blogsRouter.post('/', async (request, response, next) => {
   const body = request.body
+  const token = getTokenFrom(request)
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({error: 'token missing or invalid'})
+  }
+  const user = await User.findById(decodedToken.id)
+  // const user = await User.findById(body.userId)
+
   const blog = new Blog({
     author: body.author,
     title: body.title,
     url: body.url,
-    likes: body.likes
+    likes: body.likes,
+    user: user._id
   })
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
 
   response.json(savedBlog.toJSON())
 
@@ -37,13 +58,25 @@ blogsRouter.post('/', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const token = getTokenFrom(request)
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({error: 'token missing or invalid'})
+  }
 
   const blog = await Blog.findById(request.params.id)
-  if(!blog) return response.status(404).end()
+  if (!blog) return response.status(404).end()
+  const correctUser = blog.user.toString() === decodedToken.id.toString()
 
+  if (!correctUser) {
+    return response.status(400).status('Invalid user')
+  }
   await Blog.findByIdAndRemove(request.params.id)
-
+  const user = await User.findById(decodedToken.id)
+  user.blogs = user.blogs.filter(blog => blog.toString() !== request.params.id)
+  await user.save()
   return response.status(204).end()
+
 
 })
